Update main.py 7/22/15
Updated main.py with initial pages for database overview and table overview. Currently, the database overview lists the names of all databases, and the table overview lists the table names of the selected database. Note that selecting a database with too many tables (namely information_schema) will cause an error in curses as pagination has not been factored in.